:art: move cloning/serializing to user land :rainbow:
Creation, cloning and serializing should be left to the user.
`arbre` is agnostic to the `Node` it manipulated, it only needs a specific shape.
So it make sense to let the user manage `crud` operations on nodes.

Plus cloning was bringin a big dependency. Now bundle size has decreased
significantly :tada: